backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
httpx
mistralai
python-dotenv
pytest
